+++ lib/utils/routes.dart
@@ -1,4 +1,5 @@
 class MyRoutes{
   static String loginroute ="/login";
   static String homeroute="/home";
+  static String homeDetailsroute ="/detail";
 }
