+++ lib/widgets/home_widgets/catalog_header.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+
+
+
+
+class CatalogHeader extends StatelessWidget {
+  const CatalogHeader({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+              crossAxisAlignment:CrossAxisAlignment.start,
+              children: [
+                 Text("Catalog App",
+                 style: TextStyle(
+                    fontWeight: FontWeight.bold,
+                    fontSize: 30.0,
+                    color: Color.fromARGB(255, 16, 0, 96),
+                 ),
+                ),
+                Text("Trending Products")
+              ],
+            );
+  }
+}
